Added a webcomponent to allow browsing with the UDW
